Separate integrations from core. Also separate them by bots and ingestion
