+++ src/gurubase-backend/backend/core/admin.py
@@ -2,7 +2,7 @@ from django.conf import settings
 from django.contrib import admin
 from core.exceptions import ValidationError
 from core.models import (APIKey,
-                         Binge, CrawlState, Integration, 
+                         Binge, CrawlState, 
                          LLMEval, 
                          LinkReference, 
                          LinkValidity, 
@@ -19,8 +19,6 @@ from core.models import (APIKey,
                          SummaryQuestionGeneration, 
                          Settings, 
                          LLMEvalResult, 
-                         Thread, 
-                         WidgetId,
                          GithubFile,
                          GuruCreationForm)
 from django.utils.html import format_html
@@ -413,14 +411,6 @@ class BingeAdmin(admin.ModelAdmin):
         return None
     owner_link.short_description = 'Owner'
 
-@admin.register(WidgetId)
-class WidgetIdAdmin(admin.ModelAdmin):
-    list_display = ['id', 'guru_type', 'key', 'domain_url', 'domain', 'date_created']
-    search_fields = ['id', 'guru_type__slug', 'key', 'domain_url']
-    list_filter = ('guru_type__slug', )
-    ordering = ('-id',)
-    raw_id_fields = ('guru_type',)
-
 
 @admin.register(GithubFile)
 class GitHubFileAdmin(admin.ModelAdmin):
@@ -447,22 +437,6 @@ class APIKeyAdmin(admin.ModelAdmin):
     raw_id_fields = ('user',)
 
 
-@admin.register(Integration)
-class IntegrationAdmin(admin.ModelAdmin):
-    list_display = ['id', 'guru_type', 'type', 'workspace_name', 'date_created', 'date_updated']
-    list_filter = ('guru_type__slug', 'type')
-    search_fields = ['id', 'guru_type__slug', 'type']
-    ordering = ('-id',)
-    raw_id_fields = ('guru_type',)
-
-
-@admin.register(Thread)
-class ThreadAdmin(admin.ModelAdmin):
-    list_display = ['id', 'binge', 'integration', 'thread_id', 'date_created', 'date_updated']
-    search_fields = ['id', 'binge__id', 'integration__id', 'thread_id']
-    list_filter = ('binge__guru_type__slug', 'integration')
-    ordering = ('-id',)
-    raw_id_fields = ('binge', 'integration')
 
 @admin.register(CrawlState)
 class CrawlStateAdmin(admin.ModelAdmin):
