+++ src/gurubase-backend/backend/core/views.py
@@ -1,15 +1,10 @@
-from asgiref.sync import sync_to_async
-from rest_framework.test import APIRequestFactory
-from slack_sdk.errors import SlackApiError
 import json
 import logging
-import aiohttp
 import random
 import string
 import time
 from datetime import UTC, datetime, timedelta
 from typing import Generator
-import re
 from accounts.models import User
 from django.conf import settings
 from django.core.cache import caches
@@ -17,18 +12,18 @@ from django.core.exceptions import ValidationError
 from django.db.models import Q
 from django.http import StreamingHttpResponse
 from django.views.decorators.csrf import csrf_exempt
-from slack_sdk import WebClient
-from core.github.exceptions import GithubAppHandlerError
-from core.requester import ConfluenceRequester, GeminiRequester, OpenAIRequester, OllamaRequester
+from integrations.bots.github.app_handler import GithubAppHandler
+from integrations.models import WidgetId
+from core.requester import GeminiRequester, OpenAIRequester, OllamaRequester, ZendeskRequester
 from core.data_sources import CrawlService, YouTubeService
 from core.serializers import WidgetIdSerializer, BingeSerializer, DataSourceSerializer, GuruTypeSerializer, GuruTypeInternalSerializer, QuestionCopySerializer, FeaturedDataSourceSerializer, APIKeySerializer, DataSourceAPISerializer, SettingsSerializer
-from core.auth import auth, follow_up_examples_auth, jwt_auth, combined_auth, stream_combined_auth, api_key_auth
-from core.gcp import replace_media_root_with_base_url, replace_media_root_with_nginx_base_url
-from core.models import CrawlState, FeaturedDataSource, Question, ContentPageStatistics, Settings, WidgetId, Binge, DataSource, GuruType, Integration, Thread, APIKey, GuruCreationForm
+from core.auth import auth, follow_up_examples_auth, jwt_auth, combined_auth, stream_combined_auth, api_key_auth, widget_id_auth
+from core.gcp import replace_media_root_with_nginx_base_url
+from core.models import CrawlState, FeaturedDataSource, Question, ContentPageStatistics, Settings, Binge, DataSource, GuruType, APIKey, GuruCreationForm
 from accounts.models import User
 from core.utils import (
     # Authentication & validation
-    check_binge_auth, create_fresh_binge, decode_guru_slug, encode_guru_slug, generate_jwt, get_base_url, validate_binge_follow_up,
+    APIAskResponse, APIType, api_ask, check_binge_auth, validate_binge_follow_up,
     validate_guru_type, validate_image, 
     
     # Question & answer handling
@@ -51,88 +46,16 @@ from rest_framework.decorators import api_view, parser_classes
 from rest_framework.parsers import MultiPartParser, FormParser
 from rest_framework.response import Response
 from rest_framework import status
-from core.integrations.helpers import IntegrationError, cleanup_title, get_trust_score_emoji, strip_first_header, NotEnoughData, NotRelated
-from core.integrations.factory import IntegrationFactory
 from rest_framework.decorators import api_view, parser_classes, throttle_classes
 from rest_framework.pagination import PageNumberPagination
 from rest_framework.parsers import FormParser, MultiPartParser
 from rest_framework.response import Response
-
-from core.auth import (
-    api_key_auth,
-    auth,
-    combined_auth,
-    follow_up_examples_auth,
-    jwt_auth,
-    stream_combined_auth,
-    widget_id_auth,
-)
-from core.exceptions import NotFoundError, PermissionError
-from core.gcp import replace_media_root_with_nginx_base_url
-from core.guru_types import (
-    get_auth0_user,
-    get_guru_type_object,
-    get_guru_type_object_by_maintainer,
-    get_guru_types,
-)
 from core.handlers.response_handlers import (
     APIResponseHandler,
     DataSourceResponseHandler,
-    WidgetResponseHandler,
-)
-from core.models import (
-    APIKey,
-    Binge,
-    ContentPageStatistics,
-    DataSource,
-    FeaturedDataSource,
-    GuruType,
-    Question,
-    WidgetId,
-    Integration,
-)
-from core.requester import GeminiRequester, JiraRequester, ZendeskRequester
-from core.serializers import (
-    APIKeySerializer,
-    BingeSerializer,
-    DataSourceAPISerializer,
-    DataSourceSerializer,
-    FeaturedDataSourceSerializer,
-    GuruTypeInternalSerializer,
-    GuruTypeSerializer,
-    QuestionCopySerializer,
-    WidgetIdSerializer,
 )
 from core.services.data_source_service import DataSourceService
 from core.throttling import ConcurrencyThrottleApiKey
-from core.utils import (
-    APIAskResponse,
-    APIType,
-    api_ask,
-    check_binge_auth,
-    create_binge_helper,
-    create_custom_guru_type_slug,
-    create_guru_type_object,
-    format_date_updated,
-    format_references,
-    format_trust_score,
-    generate_jwt,
-    generate_og_image,
-    get_question_summary,
-    handle_failed_root_reanswer,
-    is_question_dirty,
-    search_question,
-    stream_and_save,
-    stream_question_answer,
-    upload_image_to_storage,
-    validate_binge_follow_up,
-    validate_guru_type,
-    validate_image,
-)
-from core.integrations.rest_commands import GetIntegrationCommand, DeleteIntegrationCommand, CreateIntegrationCommand
-from .github.event_handler import GitHubEventFactory, GitHubEventHandler
-from .github.app_handler import GithubAppHandler
-
 
 logger = logging.getLogger(__name__)
 
@@ -1439,104 +1362,6 @@ def health_check(request):
     return Response({'status': 'healthy'}, status=status.HTTP_200_OK)
 
 
-@api_view(['POST'])
-@widget_id_auth
-def ask_widget(request):
-    """
-    Widget endpoint for answering questions.
-    Supports both streaming and non-streaming responses.
-    """
-    # Initialize response handler
-    response_handler = WidgetResponseHandler()
-    
-    # Get guru type
-    guru_type_object = request.guru_type
-    
-    # Get request parameters
-    question = request.data.get('question')
-    binge_id = request.data.get('binge_id')
-    parent_slug = request.data.get('parent_slug')
-    fetch_existing = request.data.get('fetch_existing', False)
-
-    # Initialize with default values
-    binge = None
-    parent = None
-
-    # Handle binge if provided
-    if binge_id:
-        try:
-            binge = Binge.objects.get(id=binge_id)
-        except Binge.DoesNotExist:
-            return response_handler.handle_error_response("Binge not found")
-
-        if not parent_slug:
-            return response_handler.handle_error_response("Parent question slug is required")
-
-        try:
-            parent = search_question(
-                None, 
-                guru_type_object, 
-                binge, 
-                parent_slug, 
-                will_check_binge_auth=False,
-                only_widget=True
-            )
-        except Exception as e:
-            return response_handler.handle_error_response("Parent question does not exist")
-
-        if not parent:
-            return response_handler.handle_error_response("Parent question does not exist")
-
-    # Set fetch_existing to True for non-binge questions
-    # if not fetch_existing and not binge:
-    #     fetch_existing = True
-
-
-    # Get widget response
-    widget_response = api_ask(
-        question=question,
-        guru_type=guru_type_object,
-        binge=binge,
-        parent=parent,
-        fetch_existing=fetch_existing,
-        api_type=APIType.WIDGET,
-        user=None
-    )
-    
-    # Handle error case
-    if widget_response.error:
-        return response_handler.handle_error_response(widget_response.error)
-    
-    # Handle existing question case
-    if widget_response.is_existing:
-        return Response(response_handler.format_question_response(widget_response.question_obj))
-    
-    # Handle streaming case
-    if isinstance(widget_response.content, Generator):
-        return response_handler.handle_stream_response(widget_response.content)
-    
-    # Handle any other cases
-    return response_handler.handle_non_stream_response(widget_response.content)
-
-@api_view(['POST'])
-@widget_id_auth
-def widget_create_binge(request):
-    guru_type_object = request.guru_type
-    root_slug = request.data.get('root_slug')
-
-    if not root_slug:
-        return Response({'msg': 'Root question slug is required'}, status=status.HTTP_400_BAD_REQUEST)
-
-    try:
-        root_question = Question.objects.get(slug=root_slug, guru_type=guru_type_object, binge=None)
-    except Question.DoesNotExist:
-        return Response({'msg': 'Root question does not exist'}, status=status.HTTP_400_BAD_REQUEST)
-
-    binge = create_binge_helper(guru_type_object, None, root_question)
-
-    return Response({'id': str(binge.id), 'root_slug': root_slug}, status=status.HTTP_200_OK)
-
-
 @api_view(['POST', 'DELETE'])
 @jwt_auth
 def manage_widget_ids(request, guru_type):
@@ -1566,19 +1391,6 @@ def manage_widget_ids(request, guru_type):
         widget_id_obj.delete()
         return Response({'msg': 'Widget ID deleted successfully'}, status=status.HTTP_200_OK)
 
-
-@api_view(['GET'])
-@widget_id_auth
-def get_guru_visuals(request):
-    guru_type = request.guru_type
-    response = {
-        'colors': guru_type.colors,
-        'icon_url': replace_media_root_with_base_url(guru_type.icon_url),
-        'name': guru_type.name,
-        'slug': guru_type.slug,
-    }
-
-    return Response(response, status=status.HTTP_200_OK)
 
 
 @parser_classes([MultiPartParser, FormParser])
@@ -1666,6 +1478,7 @@ def api_answer(request, guru_type):
 
     github_comments = None
     if github_api_url:
+        # TODO: Fix this
         github_handler = GithubAppHandler(request.integration)
         github_issue = github_handler.get_issue(github_api_url, request.external_id)
         github_comments = github_handler.get_issue_comments(github_api_url, request.external_id)
@@ -1789,858 +1602,6 @@ def api_reindex_data_sources(request, guru_type):
         return response_handler.handle_error_response(str(e))
     except Exception as e:
         return response_handler.handle_error_response(f'Unexpected error: {str(e)}')
-
-
-@api_view(['GET'])
-def create_integration(request):
-    # OAuth flow
-    code = request.query_params.get('code')
-    state = request.query_params.get('state')
-
-    if not state:
-        return Response({
-            'msg': 'Missing required parameters'
-        }, status=status.HTTP_400_BAD_REQUEST)
-
-    try:
-        # Decode the state parameter
-        state_json = json.loads(state)
-        integration_type = state_json.get('type')
-        guru_type_slug = state_json.get('guru_type')
-        encoded_guru_slug = state_json.get('encoded_guru_slug')
-        installation_id = state_json.get('installation_id')
-
-        if not all([integration_type, guru_type_slug, encoded_guru_slug]):
-            return Response({
-                'msg': 'Invalid state parameter'
-            }, status=status.HTTP_400_BAD_REQUEST)
-
-        decoded_guru_slug = decode_guru_slug(encoded_guru_slug)
-        if not decoded_guru_slug or decoded_guru_slug != guru_type_slug:
-            return Response({
-                'msg': 'Invalid state parameter'
-            }, status=status.HTTP_400_BAD_REQUEST)                    
-
-    except Exception as e:
-        logger.error(f"Error creating integration: {e}", exc_info=True)
-        return Response({
-            'msg': 'There has been an error while creating the integration. Please try again. If the problem persists, please contact support.'
-        }, status=status.HTTP_400_BAD_REQUEST)
-
-    try:
-        guru_type = GuruType.objects.get(slug=guru_type_slug)
-    except GuruType.DoesNotExist:
-        return Response({
-            'msg': 'Invalid guru type'
-        }, status=status.HTTP_400_BAD_REQUEST)
-
-    try:
-        strategy = IntegrationFactory.get_strategy(integration_type)
-        if integration_type.upper() == 'GITHUB':
-            if not installation_id:
-                return Response({
-                    'msg': 'Missing installation_id for GitHub integration'
-                }, status=status.HTTP_400_BAD_REQUEST)
-            integration = strategy.create_integration(installation_id, guru_type)
-        else:
-            if not code:
-                return Response({
-                    'msg': 'Missing code parameter'
-                }, status=status.HTTP_400_BAD_REQUEST)
-            integration = strategy.create_integration(code, guru_type)
-    except IntegrationError as e:
-        return Response({
-            'msg': str(e)
-        }, status=status.HTTP_400_BAD_REQUEST)
-    except Exception as e:
-        logger.error(f"Error creating integration: {e}", exc_info=True)
-        return Response({
-            'msg': 'There has been an error while creating the integration. Please try again. If the problem persists, please contact support.'
-        }, status=status.HTTP_400_BAD_REQUEST)
-
-    return Response({
-        'id': integration.id,
-        'type': integration.type,
-        'guru_type': guru_type.slug,
-        'channels': integration.channels
-    })
-
-
-
-@api_view(['GET', 'DELETE', 'POST'])
-@jwt_auth
-def manage_integration(request, guru_type, integration_type):
-    """
-    GET: Get integration details for a specific guru type and integration type.
-    DELETE: Delete an integration and invalidate its OAuth token.
-    POST: Create a new integration.
-    """
-    try:
-        guru_type_object = get_guru_type_object_by_maintainer(guru_type, request)
-    except PermissionError:
-        return Response({'msg': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
-    except NotFoundError:
-        return Response({'msg': f'Guru type {guru_type} not found'}, status=status.HTTP_404_NOT_FOUND)
-        
-    # Validate integration type
-    if integration_type not in [choice.value for choice in Integration.Type]:
-        return Response({'msg': f'Invalid integration type: {integration_type}'}, status=status.HTTP_400_BAD_REQUEST)
-    
-    try:
-        if request.method in ['GET', 'DELETE']:
-            integration = Integration.objects.get(
-                guru_type=guru_type_object,
-                type=integration_type
-            )
-        else:
-            integration = None
-
-        if request.method == 'GET':
-            command = GetIntegrationCommand(integration)
-            return command.execute()
-        elif request.method == 'DELETE':
-            command = DeleteIntegrationCommand(integration, guru_type_object)
-            return command.execute()
-        elif request.method == 'POST':
-            command = CreateIntegrationCommand(guru_type_object, integration_type, request.data)
-            return command.execute()
-            
-    except Integration.DoesNotExist:
-        if request.method == 'GET':
-            return Response({"encoded_guru_slug": encode_guru_slug(guru_type_object.slug)}, status=status.HTTP_202_ACCEPTED)
-        else:
-            return Response({'msg': 'Integration not found'}, status=status.HTTP_404_NOT_FOUND)
-    except Exception as e:
-        logger.error(f"Error in manage_integration: {e}", exc_info=True)
-        return Response({'msg': 'Internal server error'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-@api_view(['GET', 'POST'])
-@jwt_auth
-def manage_channels(request, guru_type, integration_type):
-    """Get or update channels for a specific integration type of a guru type."""
-    try:
-        guru_type_object = get_guru_type_object_by_maintainer(guru_type, request)
-    except PermissionError:
-        return Response({'msg': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
-    except NotFoundError:
-        return Response({'msg': f'Guru type {guru_type} not found'}, status=status.HTTP_404_NOT_FOUND)
-        
-    # Validate integration type
-    if integration_type not in [choice.value for choice in Integration.Type]:
-        return Response({'msg': f'Invalid integration type: {integration_type}'}, status=status.HTTP_400_BAD_REQUEST)
-    
-    try:
-        integration = Integration.objects.get(
-            guru_type=guru_type_object,
-            type=integration_type
-        )
-    except Integration.DoesNotExist:
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
-    if request.method == 'POST':
-        try:
-            channels = request.data.get('channels', [])
-            integration.channels = channels
-            integration.save()
-            
-            return Response({
-                'id': integration.id,
-                'type': integration.type,
-                'guru_type': integration.guru_type.slug,
-                'channels': integration.channels
-            })
-        except Exception as e:
-            logger.error(f"Error updating channels: {e}", exc_info=True)
-            return Response({'msg': 'Internal server error'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-    try:
-        # Get channels from API
-        strategy = IntegrationFactory.get_strategy(integration_type, integration)
-        api_channels = strategy.list_channels()
-
-        processed_channels = []
-        if integration_type == 'GITHUB':
-            # Create a map of channel IDs to their allowed status from DB
-            db_channels_map = {
-                channel['id']: channel['mode']
-                for channel in integration.channels
-            }
-            # Process each API channel
-            for channel in api_channels:
-                channel['mode'] = db_channels_map.get(channel['id'], 'auto')
-                processed_channels.append(channel)
-
-            # Save the new channels, we save them in retrieval because it is set on GitHub settings. We only read what is set. We can't update them.
-            integration.channels = processed_channels
-            integration.save()
-        else:
-            # Create a map of channel IDs to their allowed status from DB
-            db_channels_map = {
-                channel['id']: channel['allowed']
-                for channel in integration.channels
-            }
-            # Process each API channel
-            for channel in api_channels:
-                # If channel exists in DB, use its allowed status
-                # Otherwise, default to False for new channels
-                channel['allowed'] = db_channels_map.get(channel['id'], False)
-                processed_channels.append(channel)
-        
-        return Response({
-            'channels': processed_channels
-        })
-    except Exception as e:
-        logger.error(f"Error listing channels: {e}", exc_info=True)
-        keyword = 'repositories' if integration_type == 'GITHUB' else 'channels'
-        return Response({'msg': f'Error listing {keyword}. Please make sure the integration is valid.'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-def get_or_create_thread_binge(thread_id: str, integration: Integration) -> tuple[Thread, Binge]:
-    """Get or create a thread and its associated binge."""
-    try:
-        thread = Thread.objects.get(thread_id=thread_id, integration=integration)
-        return thread, thread.binge
-    except Thread.DoesNotExist:
-        # Create new binge without a root question
-        binge = create_fresh_binge(integration.guru_type, None)
-        thread = Thread.objects.create(
-            thread_id=thread_id,
-            binge=binge,
-            integration=integration
-        )
-        return thread, binge
-
-def strip_first_header(content: str) -> str:
-    """Remove the first header (starting with # and ending with newline) from content."""
-    if content.startswith('#'):
-        # Find the first newline
-        newline_index = content.find('\n')
-        if newline_index != -1:
-            # Return content after the newline
-            return content[newline_index + 1:].lstrip()
-    return content
-
-def convert_markdown_to_slack(content: str) -> str:
-    """Convert Markdown formatting to Slack formatting."""
-    # Convert markdown code blocks to Slack code blocks by removing language specifiers
-    import re
-    
-    # First remove language specifiers from code blocks
-    content = re.sub(r'```\w+', '```', content)
-    
-    # Then remove empty lines at the start and end of code blocks
-    def trim_code_block(match):
-        code_block = match.group(0)
-        lines = code_block.split('\n')
-        
-        # Find first and last non-empty lines (excluding ```)
-        start = 0
-        end = len(lines) - 1
-        
-        # Find first non-empty line after opening ```
-        for i, line in enumerate(lines):
-            if line.strip() == '```':
-                start = i + 1
-                break
-                
-        # Find last non-empty line before closing ```
-        for i in range(len(lines) - 1, -1, -1):
-            if line.strip() == '```':
-                end = i
-                break
-                
-        # Keep all lines between start and end (inclusive)
-        return '```\n' + '\n'.join(lines[start:end]) + '\n```'
-    
-    content = re.sub(r'```[\s\S]+?```', trim_code_block, content)
-    
-    # Convert markdown links [text](url) to Slack format <url|text>
-    def replace_link(match):
-        text = match.group(1)
-        url = match.group(2)
-        return f"<{url}|{text}>"
-    
-    content = re.sub(r'\[([^\]]+)\]\(([^)]+)\)', replace_link, content)
-    
-    # Convert markdown bold/italic to Slack format
-    # First handle single asterisks for italics (but not if they're part of double asterisks)
-    i = 0
-    while i < len(content):
-        if content[i:i+2] == "**":
-            i += 2
-        elif content[i] == "*":
-            # Replace single asterisk with underscore for italics
-            content = content[:i] + "_" + content[i+1:]
-        i += 1
-    
-    # Then handle double asterisks for bold
-    content = content.replace("**", "*")
-    
-    return content
-
-def format_slack_response(content: str, trust_score: int, references: list, question_url: str) -> str:
-    """Format the response with trust score and references for Slack.
-    Using Slack's formatting syntax:
-    *bold*
-    _italic_
-    ~strikethrough~
-    `code`
-    ```preformatted```
-    >blockquote
-    <url|text> for links
-    """
-    # Strip header from content
-    content = strip_first_header(content)
-    
-    # Convert markdown to slack formatting
-    content = convert_markdown_to_slack(content)
-    
-    formatted_msg = [content]
-    
-    # Add trust score with emoji
-    trust_emoji = get_trust_score_emoji(trust_score)
-    formatted_msg.append(f"\n---------\n_*Trust Score*: {trust_emoji} {trust_score}_%")
-    
-    # Add references if they exist
-    if references:
-        formatted_msg.append("\n_*Sources*_:")
-        for ref in references:
-            # First remove Slack-style emoji codes with adjacent spaces
-            clean_title = cleanup_title(ref['title'])
-            
-            formatted_msg.append(f"\n• _<{ref['link']}|{clean_title}>_")
-    
-    # Add frontend link if it exists
-    if question_url:
-        formatted_msg.append(f"\n:eyes: _<{question_url}|View on Gurubase for a better UX>_")
-    
-    return "\n".join(formatted_msg)
-
-async def stream_and_update_message(
-    session: aiohttp.ClientSession,
-    url: str,
-    headers: dict,
-    payload: dict,
-    client: WebClient,
-    channel_id: str,
-    message_ts: str,
-    update_interval: float = 0.5
-) -> None:
-    """Stream the response and update the Slack message periodically."""
-    last_update = time.time()
-    current_content = ""
-    
-    try:
-        # Create request using APIRequestFactory
-        factory = APIRequestFactory()
-        guru_type = payload.get('guru_type')
-        
-        request = factory.post(
-            f'/api/v1/{guru_type}/answer/',
-            payload,
-            HTTP_X_API_KEY=headers.get('X-API-KEY'),
-            format='json'
-        )
-        
-        # Call api_answer directly in a sync context
-        response = await sync_to_async(api_answer)(request, guru_type)
-        
-        # Handle StreamingHttpResponse
-        if hasattr(response, 'streaming_content'):
-            buffer = ""
-            line_buffer = ""
-            
-            # Create an async wrapper for the generator iteration
-            @sync_to_async
-            def get_next_chunk():
-                try:
-                    return next(response.streaming_content)
-                except StopIteration:
-                    return None
-            
-            # Iterate over the generator asynchronously
-            while True:
-                chunk = await get_next_chunk()
-                if chunk is None:
-                    # Yield any remaining text in the buffer
-                    if line_buffer.strip():
-                        buffer += line_buffer
-                        # Strip header and convert markdown
-                        cleaned_content = strip_first_header(buffer)
-                        if cleaned_content.strip():
-                            formatted_content = convert_markdown_to_slack(cleaned_content)
-                            formatted_content += '\n\n:clock1: _streaming..._'
-                            try:
-                                client.chat_update(
-                                    channel=channel_id,
-                                    ts=message_ts,
-                                    text=formatted_content
-                                )
-                            except SlackApiError as e:
-                                logger.error(f"Error updating message: {e.response}", exc_info=True)
-                    break
-                    
-                if chunk:
-                    text = chunk.decode('utf-8') if isinstance(chunk, bytes) else str(chunk)
-                    line_buffer += text
-                    
-                    # Check if we have complete lines
-                    while '\n' in line_buffer:
-                        line, line_buffer = line_buffer.split('\n', 1)
-                        if line.strip():
-                            buffer += line + '\n'
-                            # Strip header and convert markdown
-                            cleaned_content = strip_first_header(buffer)
-                            if cleaned_content.strip():
-                                formatted_content = convert_markdown_to_slack(cleaned_content)
-                                formatted_content += '\n\n:clock1: _streaming..._'
-                                current_time = time.time()
-                                if current_time - last_update >= update_interval:
-                                    try:
-                                        client.chat_update(
-                                            channel=channel_id,
-                                            ts=message_ts,
-                                            text=formatted_content
-                                        )
-                                        last_update = current_time
-                                    except SlackApiError as e:
-                                        logger.error(f"Error updating message: {e.response}", exc_info=True)
-                                        client.chat_update(
-                                            channel=channel_id,
-                                            ts=message_ts,
-                                            text="❌ Failed to update message"
-                                        )
-                                        return
-    except Exception as e:
-        logger.error(f"Error in stream_and_update_message: {str(e)}", exc_info=True)
-        client.chat_update(
-            channel=channel_id,
-            ts=message_ts,
-            text="❌ An error occurred while processing your request"
-        )
-        return
-
-async def get_final_response(
-    session: aiohttp.ClientSession,
-    url: str,
-    headers: dict,
-    payload: dict,
-    client: WebClient,
-    channel_id: str,
-    message_ts: str
-) -> None:
-    """Get and send the final formatted response."""
-    try:
-        # Create request using APIRequestFactory
-        factory = APIRequestFactory()
-        guru_type = payload.get('guru_type')
-        
-        request = factory.post(
-            f'/api/v1/{guru_type}/answer/',
-            payload,
-            HTTP_X_API_KEY=headers.get('X-API-KEY'),
-            format='json'
-        )
-        
-        # Call api_answer directly
-        response = await sync_to_async(api_answer)(request, guru_type)
-        
-        # Convert response to dict if it's a Response object
-        if hasattr(response, 'data'):
-            final_response = response.data
-        else:
-            final_response = response
-
-        if 'msg' in final_response and 'doesn\'t have enough data' in final_response['msg']:
-            raise NotEnoughData(final_response['msg'])
-        elif 'msg' in final_response and 'is not related to' in final_response['msg']:
-            raise NotRelated(final_response['msg'])
-        elif 'msg' in final_response:
-            raise Exception(final_response['msg'])
-
-        trust_score = final_response.get('trust_score', 0)
-        references = final_response.get('references', [])
-        content = final_response.get('content', '')
-        question_url = final_response.get('question_url', '')
-
-        final_text = format_slack_response(content, trust_score, references, question_url)
-        if final_text.strip():  # Only update if there's content after stripping header
-            client.chat_update(
-                channel=channel_id,
-                ts=message_ts,
-                text=final_text
-            )
-    except NotEnoughData as e:
-        logger.error(f"Not enough data: {str(e)}", exc_info=True)
-        client.chat_update(
-            channel=channel_id,
-            ts=message_ts,
-            text=f"❌ {str(e)}"
-        )
-    except NotRelated as e:
-        logger.error(f"Not related to the question: {str(e)}", exc_info=True)
-        client.chat_update(
-            channel=channel_id,
-            ts=message_ts,
-            text=f"❌ {str(e)}"
-        )
-    except Exception as e:
-        logger.error(f"Error in get_final_response: {str(e)}", exc_info=True)
-        client.chat_update(
-            channel=channel_id,
-            ts=message_ts,
-            text="❌ An error occurred while processing your request"
-        )
-
-async def handle_slack_message(
-    client: WebClient,
-    integration: Integration,
-    channel_id: str,
-    thread_ts: str,
-    clean_message: str
-) -> None:
-    """Handle a single Slack message."""
-
-    try:
-        # First send a thinking message
-        thinking_response = client.chat_postMessage(
-            channel=channel_id,
-            thread_ts=thread_ts,
-            text="Thinking... 🤔"
-        )
-        
-        try:
-            # Get or create thread and binge
-            thread, binge = await sync_to_async(get_or_create_thread_binge)(thread_ts, integration)
-        except Exception as e:
-            logger.error(f"Error creating thread/binge: {str(e)}", exc_info=True)
-            client.chat_update(
-                channel=channel_id,
-                ts=thinking_response["ts"],
-                text="❌ Failed to create conversation thread"
-            )
-            return
-        
-        guru_type_slug = await sync_to_async(lambda integration: integration.guru_type.slug)(integration)
-        api_key = await sync_to_async(lambda integration: integration.api_key.key)(integration)
-        
-        try:
-            # First get streaming response
-            stream_payload = {
-                'question': clean_message,
-                'stream': True,
-                'short_answer': True,
-                'session_id': str(binge.id),
-                'guru_type': guru_type_slug
-            }
-            
-            headers = {
-                'X-API-KEY': api_key,
-                'Content-Type': 'application/json'
-            }
-            
-            async with aiohttp.ClientSession() as session:
-                await stream_and_update_message(
-                    session=session,
-                    url='',  # Not used anymore
-                    headers=headers,
-                    payload=stream_payload,
-                    client=client,
-                    channel_id=channel_id,
-                    message_ts=thinking_response["ts"]
-                )
-                
-                # Then get final formatted response
-                final_payload = {
-                    'question': clean_message,
-                    'stream': False,
-                    'short_answer': True,
-                    'fetch_existing': True,
-                    'session_id': str(binge.id),
-                    'guru_type': guru_type_slug
-                }
-                
-                await get_final_response(
-                    session=session,
-                    url='',  # Not used anymore
-                    headers=headers,
-                    payload=final_payload,
-                    client=client,
-                    channel_id=channel_id,
-                    message_ts=thinking_response["ts"]
-                )
-        except aiohttp.ClientError as e:
-            logger.error(f"Network error: {str(e)}", exc_info=True)
-            client.chat_update(
-                channel=channel_id,
-                ts=thinking_response["ts"],
-                text="❌ Network error occurred while processing your request"
-            )
-        except Exception as e:
-            logger.error(f"Error in API communication: {str(e)}", exc_info=True)
-            client.chat_update(
-                channel=channel_id,
-                ts=thinking_response["ts"],
-                text="❌ An error occurred while processing your request"
-            )
-    except SlackApiError as e:
-        logger.error(f"Slack API error: {str(e)}", exc_info=True)
-        # If we can't even send the thinking message, we can't update it later
-        try:
-            if thinking_response:
-                client.chat_update(
-                    channel=channel_id,
-                    ts=thinking_response["ts"],
-                    text="❌ Failed to process your request due to a Slack API error"
-                )
-            else:
-                client.chat_postMessage(
-                    channel=channel_id,
-                    thread_ts=thread_ts,
-                    text="❌ Failed to process your request due to a Slack API error"
-                )
-        except:
-            pass  # If this fails too, we can't do much
-    except Exception as e:
-        logger.error(f"Unexpected error in handle_slack_message: {str(e)}", exc_info=True)
-        try:
-            if thinking_response:
-                client.chat_update(
-                    channel=channel_id,
-                    ts=thinking_response["ts"],
-                    text="❌ An unexpected error occurred"
-                )
-            else:
-                client.chat_postMessage(
-                    channel=channel_id,
-                    thread_ts=thread_ts,
-                    text="❌ An unexpected error occurred"
-                )
-        except:
-            pass  # If this fails too, we can't do much
-
-async def send_channel_unauthorized_message(
-    client: WebClient,
-    channel_id: str,
-    thread_ts: str,
-    guru_slug: str
-) -> None:
-    """Send a message explaining how to authorize the channel."""
-    try:
-        base_url = await sync_to_async(get_base_url)()
-        settings_url = f"{base_url.rstrip('/')}/guru/{guru_slug}/integrations/slack"
-        message = (
-            "❌ This channel is not authorized to use the bot.\n\n"
-            f"Please visit <{settings_url}|Gurubase Settings> to configure "
-            "the bot and add this channel to the allowed channels list."
-        )
-        client.chat_postMessage(
-            channel=channel_id,
-            thread_ts=thread_ts,
-            text=message
-        )
-    except SlackApiError as e:
-        logger.error(f"Error sending unauthorized channel message: {e.response}", exc_info=True)
-
-
-@api_view(['GET', 'POST'])
-def slack_events(request):
-    """Handle Slack events including verification and message processing."""
-    from slack_sdk import WebClient
-    from slack_sdk.errors import SlackApiError
-    import asyncio
-    import threading
-    from django.core.cache import caches
-    
-    data = request.data
-    
-    # If this is a verification request, respond with the challenge parameter
-    if "challenge" in data:
-        return Response(data["challenge"], status=status.HTTP_200_OK)
-    
-    # Handle the event in a separate thread
-    if "event" in data:
-        def process_event():
-            try:
-                event = data["event"]
-                
-                # Only proceed if it's a message event and not from a bot
-                if event["type"] == "message" and "subtype" not in event and event.get("user") != event.get("bot_id"):
-                    # Get bot user ID from authorizations
-                    bot_user_id = data.get("authorizations", [{}])[0].get("user_id")
-                    user_message = event["text"]
-                    
-                    # First check if the bot is mentioned
-                    if not (bot_user_id and f"<@{bot_user_id}>" in user_message):
-                        return
-                        
-                    team_id = data.get('team_id')
-                    if not team_id:
-                        return
-                        
-                    # Try to get integration from cache first
-                    cache = caches['alternate']
-                    cache_key = f"slack_integration:{team_id}"
-                    integration = cache.get(cache_key)
-                    
-                    if not integration:
-                        try:
-                            # If not in cache, get from database
-                            integration = Integration.objects.get(type=Integration.Type.SLACK, external_id=team_id)
-                            # Set cache timeout to 0. This is because dynamic channel updates are not immediately reflected
-                            # And this may result in bad UX, and false positive bug reports
-                            cache.set(cache_key, integration, timeout=0)
-                        except Integration.DoesNotExist:
-                            logger.error(f"No integration found for team {team_id}", exc_info=True)
-                            return
-                    
-                    try:
-                        # Get the Slack client for this team
-                        client = WebClient(token=integration.access_token)
-                        
-                        channel_id = event["channel"]
-                        
-                        # Check if the current channel is allowed
-                        channels = integration.channels
-                        channel_allowed = False
-                        for channel in channels:
-                            if str(channel.get('id')) == channel_id and channel.get('allowed', False):
-                                channel_allowed = True
-                                break
-
-                        # Get thread_ts if it exists (means we're in a thread)
-                        thread_ts = event.get("thread_ts") or event.get("ts")
-                        
-                        if not channel_allowed:
-                            # Run the unauthorized message handler in the event loop
-                            loop = asyncio.new_event_loop()
-                            asyncio.set_event_loop(loop)
-                            try:
-                                loop.run_until_complete(send_channel_unauthorized_message(
-                                    client=client,
-                                    channel_id=channel_id,
-                                    thread_ts=thread_ts,
-                                    guru_slug=integration.guru_type.slug
-                                ))
-                            finally:
-                                loop.close()
-                            return
-                        
-                        # Remove the bot mention from the message
-                        clean_message = user_message.replace(f"<@{bot_user_id}>", "").strip()
-                        
-                        # Run the async handler in a new event loop
-                        loop = asyncio.new_event_loop()
-                        asyncio.set_event_loop(loop)
-                        try:
-                            loop.run_until_complete(handle_slack_message(
-                                client=client,
-                                integration=integration,
-                                channel_id=channel_id,
-                                thread_ts=thread_ts,
-                                clean_message=clean_message
-                            ))
-                        except SlackApiError as e:
-                            if e.response.data.get('msg') in ['token_expired', 'invalid_auth', 'not_authed']:
-                                try:
-                                    # Get fresh integration data from DB
-                                    integration = Integration.objects.get(id=integration.id)
-                                    # Try to refresh the token
-                                    strategy = IntegrationFactory.get_strategy(integration.type, integration)
-                                    new_token = strategy.handle_token_refresh()
-                                    
-                                    # Update cache with new integration data
-                                    cache.set(cache_key, integration, timeout=300)
-                                    
-                                    # Retry with new token
-                                    client = WebClient(token=new_token)
-                                    loop.run_until_complete(handle_slack_message(
-                                        client=client,
-                                        integration=integration,
-                                        channel_id=channel_id,
-                                        thread_ts=thread_ts,
-                                        clean_message=clean_message
-                                    ))
-                                except Exception as refresh_error:
-                                    logger.error(f"Error refreshing token: {refresh_error}", exc_info=True)
-                            else:
-                                logger.error(f"Slack API error: {e}", exc_info=True)
-                        finally:
-                            loop.close()
-                            
-                    except Exception as e:
-                        logger.error(f"Error processing Slack event: {e}", exc_info=True)
-            except Exception as e:
-                logger.error(f"Error in process_event thread: {e}", exc_info=True)
-        
-        # Start processing in a separate thread
-        thread = threading.Thread(target=process_event)
-        thread.daemon = True  # Make thread daemon so it doesn't block server shutdown
-        thread.start()
-    
-    # Return 200 immediately
-    return Response(status=200)
-
-@api_view(['POST'])
-@jwt_auth
-def send_test_message(request):
-    """Send a test message to a specific channel using the specified integration."""
-    integration_id = request.data.get('integration_id')
-    channel_id = request.data.get('channel_id')
-
-    if not integration_id or not channel_id:
-        return Response({'msg': 'Integration ID and channel ID are required'}, status=status.HTTP_400_BAD_REQUEST)
-
-    try:
-        integration = Integration.objects.get(id=integration_id)
-    except Integration.DoesNotExist:
-        return Response({'msg': 'Integration not found'}, status=status.HTTP_404_NOT_FOUND)
-
-    try:
-        # Get the appropriate strategy for the integration type
-        strategy = IntegrationFactory.get_strategy(integration.type, integration)
-        success = strategy.send_test_message(channel_id)
-        
-        if success:
-            return Response({'msg': 'Test message sent successfully'}, status=status.HTTP_200_OK)
-        else:
-            return Response({'msg': 'Failed to send test message'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-    except Exception as e:
-        logger.error(f"Error sending test message: {e}", exc_info=True)
-        return Response({'msg': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-@api_view(['GET'])
-@jwt_auth
-def list_integrations(request, guru_type):
-    """
-    GET: List all integrations for a specific guru type.
-    """
-    try:
-        guru_type_object = get_guru_type_object_by_maintainer(guru_type, request)
-    except PermissionError:
-        return Response({'msg': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
-    except NotFoundError:
-        return Response({'msg': f'Guru type {guru_type} not found'}, status=status.HTTP_404_NOT_FOUND)
-    
-    try:
-        integrations = Integration.objects.filter(guru_type=guru_type_object)
-        
-        response_data = []
-        for integration in integrations:
-            response_data.append({
-                'id': integration.id,
-                'type': integration.type,
-                'workspace_name': integration.workspace_name,
-                'external_id': integration.external_id,
-                'channels': integration.channels,
-                'date_created': integration.date_created,
-                'date_updated': integration.date_updated,
-            })
-        
-        return Response(response_data, status=status.HTTP_200_OK)
-            
-    except Exception as e:
-        logger.error(f"Error in list_integrations: {e}", exc_info=True)
-        return Response({'msg': 'Internal server error'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 @api_view(['GET', 'PUT'])
 @jwt_auth
@@ -2978,257 +1939,6 @@ def fetch_youtube_channel_api(request):
         return Response({'error': str(e)}, status=status.HTTP_400_BAD_REQUEST)
     return Response(data, status=return_status)
 
-@api_view(['GET', 'POST'])
-def github_webhook(request):
-    if request.method != 'POST':
-        return Response({'message': 'Webhook received'}, status=status.HTTP_200_OK)
-
-    body = request.body
-    data = request.data
-    installation_id = data.get('installation', {}).get('id')
-    
-    if not installation_id:
-        logger.error("No installation ID found in webhook payload")
-        return Response({'message': 'No installation ID found'}, status=status.HTTP_400_BAD_REQUEST)
-
-    # Try to get integration from cache first
-    cache = caches['alternate']
-    cache_key = f"github_integration:{installation_id}"
-    integration = cache.get(cache_key)
-    
-    if not integration:
-        try:
-            # If not in cache, get from database
-            integration = Integration.objects.get(type=Integration.Type.GITHUB, external_id=installation_id)
-            # Set cache timeout to 0. This is because dynamic updates are not immediately reflected
-            # And this may result in bad UX, and false positive bug reports
-            cache.set(cache_key, integration, timeout=0)
-        except Integration.DoesNotExist:
-            logger.error(f"No integration found for installation {installation_id}", exc_info=True)
-            return Response({'message': 'No integration found'}, status=status.HTTP_400_BAD_REQUEST)
-
-    assert integration is not None
-
-    bot_name = integration.github_bot_name
-    github_handler = GithubAppHandler(integration)
-    # Verify GitHub webhook signature
-    try:
-        signature_header = request.headers.get('x-hub-signature-256')
-        github_handler.verify_signature(body, signature_header)
-    except GithubAppHandlerError as e:
-        logger.error(f"GitHub webhook signature verification failed: {e}")
-        return Response({'message': 'Invalid signature'}, status=status.HTTP_403_FORBIDDEN)
-
-    try:
-        # Get event type and validate
-        event_type = GitHubEventHandler.find_github_event_type(data)
-        if not GitHubEventHandler.is_supported_event(event_type):
-            return Response({'message': 'Webhook received'}, status=status.HTTP_200_OK)
-
-        # Get the appropriate event handler
-        event_handler = GitHubEventFactory.get_handler(event_type, integration, github_handler)
-        
-        # Extract event data
-        event_data = event_handler.extract_event_data(data)
-        
-        # Find the appropriate channel
-        channel = None
-        for channel_itr in integration.channels:
-            if channel_itr['name'] == event_data['repository_name']:
-                channel = channel_itr
-                break
-
-        if not channel:
-            logger.error(f"No channel found for repository {event_data['repository_name']}", exc_info=True)
-            return Response({'message': 'No channel found'}, status=status.HTTP_400_BAD_REQUEST)
-
-        # Check if we should answer
-        if not github_handler.will_answer(event_data['body'], bot_name, event_type, channel['mode']):
-            return Response({'message': 'Webhook received'}, status=status.HTTP_200_OK)
-
-        # Create a new binge
-        binge = create_fresh_binge(integration.guru_type, None)
-        guru_type = integration.guru_type.slug
-        
-        # Create request using APIRequestFactory
-        factory = APIRequestFactory()
-        
-        # Prepare payload for the API
-        payload = {
-            'question': github_handler.cleanup_user_question(event_data['body'], bot_name),
-            'stream': False,
-            'short_answer': True,
-            'fetch_existing': False,
-            'session_id': binge.id
-        }
-
-        if event_data['api_url']:
-            payload['github_api_url'] = event_data['api_url']
-        
-        # Create request with API key from integration
-        request = factory.post(
-            f'/api/v1/{guru_type}/answer/',
-            payload,
-            HTTP_X_API_KEY=integration.api_key.key,
-            format='json'
-        )
-        
-        # Call api_answer directly
-        response = api_answer(request, guru_type)
-        
-        # Handle the response using the event handler
-        event_handler.handle_response(response, event_data, bot_name)
-
-    except Exception as e:
-        logger.error(f"Error processing GitHub webhook: {e}", exc_info=True)
-        return Response({'message': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-            
-    return Response({'message': 'Webhook received'}, status=status.HTTP_200_OK)
-
-@api_view(['POST'])
-@jwt_auth
-def list_jira_issues(request, integration_id):
-    """
-    List Jira issues for a specific integration using a JQL query.
-    Requires integration_id in the path and accepts 'jql' as a query parameter.
-    """
-    try:
-        integration = Integration.objects.get(id=integration_id)
-    except Integration.DoesNotExist:
-        return Response({'msg': 'Integration not found'}, status=status.HTTP_404_NOT_FOUND)
-
-    # Validate user permission (e.g., maintainer or admin)
-    try:
-        get_guru_type_object_by_maintainer(integration.guru_type.slug, request)
-    except PermissionError:
-        return Response({'msg': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
-    except NotFoundError:
-        # This shouldn't happen if integration exists, but good practice
-        return Response({'msg': 'Associated Guru type not found'}, status=status.HTTP_404_NOT_FOUND)
-
-    # Validate integration type
-    if integration.type != Integration.Type.JIRA:
-        return Response({'msg': 'This integration is not a Jira integration.'}, status=status.HTTP_400_BAD_REQUEST)
-
-    # Get JQL from query params, default if not provided
-    jql_query = request.data.get('jql')
-    if not jql_query:
-        jql_query = 'ORDER BY created DESC'
-
-    try:
-        jira_requester = JiraRequester(integration)
-        issues = jira_requester.list_issues(jql_query=jql_query)
-        return Response({'issues': issues, 'issue_count': len(issues)}, status=status.HTTP_200_OK)
-    except ValueError as e:
-        # Handle specific errors from JiraRequester
-        error_str = str(e)
-        if "Invalid Jira credentials" in error_str:
-             return Response({'msg': 'Invalid Jira credentials.'}, status=status.HTTP_401_UNAUTHORIZED)
-        elif "Jira API access forbidden" in error_str:
-             return Response({'msg': 'Jira API access forbidden. Check user permissions or API key scope.'}, status=status.HTTP_403_FORBIDDEN)
-        else:
-             logger.error(f"Error listing Jira issues for integration {integration_id}: {e}", exc_info=True)
-             return Response({'msg': f'Failed to list Jira issues: {error_str}'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-    except Exception as e:
-        logger.error(f"Unexpected error listing Jira issues for integration {integration_id}: {e}", exc_info=True)
-        return Response({'msg': 'An unexpected error occurred.'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-@api_view(['GET'])
-@jwt_auth
-def list_zendesk_tickets(request, integration_id):
-    """
-    List Zendesk tickets for a specific integration.
-    Requires integration_id in the path.
-    """
-    try:
-        integration = Integration.objects.get(id=integration_id)
-    except Integration.DoesNotExist:
-        return Response({'msg': 'Integration not found'}, status=status.HTTP_404_NOT_FOUND)
-
-    # Validate user permission (e.g., maintainer or admin)
-    try:
-        get_guru_type_object_by_maintainer(integration.guru_type.slug, request)
-    except PermissionError:
-        return Response({'msg': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
-    except NotFoundError:
-        return Response({'msg': 'Associated Guru type not found'}, status=status.HTTP_404_NOT_FOUND)
-
-    # Validate integration type
-    if integration.type != Integration.Type.ZENDESK:
-        return Response({'msg': 'This integration is not a Zendesk integration.'}, status=status.HTTP_400_BAD_REQUEST)
-
-    try:
-        zendesk_requester = ZendeskRequester(integration)
-        tickets = zendesk_requester.list_tickets()
-        return Response({'tickets': tickets, 'ticket_count': len(tickets)}, status=status.HTTP_200_OK)
-    except ValueError as e:
-        # Handle specific errors from ZendeskRequester
-        error_str = str(e)
-        if "Zendesk credentials" in error_str:
-             return Response({'msg': 'Missing or invalid Zendesk credentials.'}, status=status.HTTP_401_UNAUTHORIZED)
-        elif "Authentication failed" in error_str:
-             return Response({'msg': 'Zendesk authentication failed. Check email and API token.'}, status=status.HTTP_401_UNAUTHORIZED)
-        elif "Permission denied" in error_str:
-             return Response({'msg': 'Zendesk permission denied. Check API token scope.'}, status=status.HTTP_403_FORBIDDEN)
-        elif "Resource not found" in error_str or "invalid Zendesk domain" in error_str:
-            return Response({'msg': 'Zendesk resource not found or invalid domain.'}, status=status.HTTP_404_NOT_FOUND)
-        elif "rate limit exceeded" in error_str:
-            return Response({'msg': 'Zendesk API rate limit exceeded.'}, status=status.HTTP_429_TOO_MANY_REQUESTS)
-        else:
-             logger.error(f"Error listing Zendesk tickets for integration {integration_id}: {e}", exc_info=True)
-             return Response({'msg': f'Failed to list Zendesk tickets: {error_str}'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-    except Exception as e:
-        logger.error(f"Unexpected error listing Zendesk tickets for integration {integration_id}: {e}", exc_info=True)
-        return Response({'msg': 'An unexpected error occurred.'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-@api_view(['GET'])
-@jwt_auth
-def list_zendesk_articles(request, integration_id):
-    """
-    List Zendesk help center articles for a specific integration.
-    Requires integration_id in the path.
-    """
-    try:
-        integration = Integration.objects.get(id=integration_id)
-    except Integration.DoesNotExist:
-        return Response({'msg': 'Integration not found'}, status=status.HTTP_404_NOT_FOUND)
-
-    # Validate user permission (e.g., maintainer or admin)
-    try:
-        get_guru_type_object_by_maintainer(integration.guru_type.slug, request)
-    except PermissionError:
-        return Response({'msg': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
-    except NotFoundError:
-        return Response({'msg': 'Associated Guru type not found'}, status=status.HTTP_404_NOT_FOUND)
-
-    # Validate integration type
-    if integration.type != Integration.Type.ZENDESK:
-        return Response({'msg': 'This integration is not a Zendesk integration.'}, status=status.HTTP_400_BAD_REQUEST)
-
-    try:
-        zendesk_requester = ZendeskRequester(integration)
-        articles = zendesk_requester.list_articles()
-        return Response({'articles': articles, 'article_count': len(articles)}, status=status.HTTP_200_OK)
-    except ValueError as e:
-        # Handle specific errors from ZendeskRequester
-        error_str = str(e)
-        if "Zendesk credentials" in error_str:
-             return Response({'msg': 'Missing or invalid Zendesk credentials.'}, status=status.HTTP_401_UNAUTHORIZED)
-        elif "Authentication failed" in error_str:
-             return Response({'msg': 'Zendesk authentication failed. Check email and API token.'}, status=status.HTTP_401_UNAUTHORIZED)
-        elif "Permission denied" in error_str:
-             return Response({'msg': 'Zendesk permission denied. Check API token scope.'}, status=status.HTTP_403_FORBIDDEN)
-        elif "Resource not found" in error_str or "invalid Zendesk domain" in error_str:
-            return Response({'msg': 'Zendesk resource not found or invalid domain.'}, status=status.HTTP_404_NOT_FOUND)
-        elif "rate limit exceeded" in error_str:
-            return Response({'msg': 'Zendesk API rate limit exceeded.'}, status=status.HTTP_429_TOO_MANY_REQUESTS)
-        else:
-             logger.error(f"Error listing Zendesk articles for integration {integration_id}: {e}", exc_info=True)
-             return Response({'msg': f'Failed to list Zendesk articles: {error_str}'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-    except Exception as e:
-        logger.error(f"Unexpected error listing Zendesk articles for integration {integration_id}: {e}", exc_info=True)
-        return Response({'msg': 'An unexpected error occurred.'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
 
 @api_view(['POST'])
 @jwt_auth
@@ -3253,58 +1963,3 @@ def validate_ollama_url(request):
         'is_valid': True,
         'models': models
     })
-
-@api_view(['POST'])
-@jwt_auth
-def list_confluence_pages(request, integration_id):
-    """
-    List Confluence pages with optional filtering.
-    
-    This endpoint handles both spaces and pages, functioning as a combined endpoint:
-    - If cql parameter is provided, returns pages matching the query
-    - If no parameters are provided, returns all pages
-    
-    Requires integration_id in the path.
-    Optional POST body parameters:
-    - cql: Confluence Query Language for filtering pages
-    """
-    try:
-        integration = Integration.objects.get(id=integration_id)
-    except Integration.DoesNotExist:
-        return Response({'msg': 'Integration not found'}, status=status.HTTP_404_NOT_FOUND)
-
-    # Validate user permission (e.g., maintainer or admin)
-    try:
-        get_guru_type_object_by_maintainer(integration.guru_type.slug, request)
-    except PermissionError:
-        return Response({'msg': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
-    except NotFoundError:
-        # This shouldn't happen if integration exists, but good practice
-        return Response({'msg': 'Associated Guru type not found'}, status=status.HTTP_404_NOT_FOUND)
-
-    # Validate integration type
-    if integration.type != Integration.Type.CONFLUENCE:
-        return Response({'msg': 'This integration is not a Confluence integration.'}, status=status.HTTP_400_BAD_REQUEST)
-
-    # Extract parameters from request data with defaults
-    cql = request.data.get('query')
-
-    try:
-        confluence_requester = ConfluenceRequester(integration)
-        result = confluence_requester.list_pages(
-            cql=cql 
-        )
-        return Response(result, status=status.HTTP_200_OK)
-    except ValueError as e:
-        # Handle specific errors from ConfluenceRequester
-        error_str = str(e)
-        if "Invalid Confluence credentials" in error_str:
-             return Response({'msg': 'Invalid Confluence credentials.'}, status=status.HTTP_401_UNAUTHORIZED)
-        elif "Confluence API access forbidden" in error_str:
-             return Response({'msg': 'Confluence API access forbidden. Check user permissions or API token scope.'}, status=status.HTTP_403_FORBIDDEN)
-        else:
-             logger.error(f"Error listing Confluence content for integration {integration_id}: {e}", exc_info=True)
-             return Response({'msg': f'Failed to list Confluence pages: {error_str}'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-    except Exception as e:
-        logger.error(f"Unexpected error listing Confluence content for integration {integration_id}: {e}", exc_info=True)
-        return Response({'msg': 'An unexpected error occurred.'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
