+++ src/gurubase-backend/backend/integrations/bots/github/pem/github.pem
@@ -1,0 +1,27 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIIEowIBAAKCAQEAuQruTuS8b+0jS0T04rvoQ0vWj2vNabSTITaeJkH4QsgVrG8J
+gV2wd2+Ouj9qsJmOPSIinr9ferGUTU9jYXFAeqoySTtXVKpxaEn1w24637eeg5Jx
+XREz2AaBDu2P+TayX7d9yX+VfbJVMidRqzD1mEQr65er9+T6EqX1IOJDGxuQqBtS
+qImGkCd1HdhJPoBnnAChePTnDSAu5b6jMH5bYdjDqfA73FyZu4fQAeHzwR84h0+F
+KrPEa18X7km7f7x+0jWycjYnCRZK587kNxBNkCl08kXWPgVOiFBSbbFpsW9w+UZC
+NYKuSuOPwaZySfsIFdbldezSg5gUvfKQqB9dJQIDAQABAoIBACRkUvPvJmQ41h0i
+harl28sIaj0wrWhhnCIHFNlfVg0EPsb32322yFLk/MQCXuA2gvPg98AS0RtXO1xt
+BZWPaWf2ewdOBWHgV4d0LEJwNmHzQdiMf4EXNgQBW3ZtIEiaCkCq3LE5R1Q+M2U2
+uGgbq+FzXtPps2eAf/IU832xIhN9ZX4klQ2GERQWtfZgmGEh8F46tw0VNBjjCTeh
+r3SwLR9euQfRuDhOusB7NIiQKlXFAuVJulaQ6ikAM3MMfEFJHY8SNJQ8ZeTSrBa2
+czgbm3U6fvWVG4qvsJS9FesL4+J20POXhmWkBnVkdHbHXF9PRMK8xzxgEKwQcV5R
+zJTbmm0CgYEA6rPEyqE9AlHTHjcaDWZOWo0ADjNrVfxtHaoNQHtzHwD7lqKDGFNo
+lZaIxOCmSpZ9RqothRblodxW9jRRJWJwSzp9zkP1RubOcCgsuLVG3+S5QC8RYQ0I
+sqy3AfRNCONFsOOtsLcREZl6lYppOB6G7mVz8d4LCL5geI0AP2z5H38CgYEAydWM
+8vZ+TGOD3jLnA9v2g9RUG92sqLgNnZqTO8U9F81A3+oTTXuWTnvA/p8VghRhYWHl
+r+iuFVUWYjaL0QybNxrAf5xym21eBamYjuKEiT3DBsfZ7L6Enic9VzmADBFDNHyh
+SbWw8x+qSULw+XJEsEgj/Wbqmxy5+Zs+fVnkVVsCgYEAsPSBN1WCLjGLP4jD975e
+7R7KH+rczqPXkToZa7in/4Q6LESzGrAodkBdNjM6rZnxMko0r9enovYCCvc/BzTE
+/bNIDzBlyeNTWnSv40OqADzu/1l9fPLaz/wwrySGwGC6treyyT39fIt/RShqVImP
+qkaLuJx1LU58ZefJfAipe/kCgYAouELVidF1S+aA37gn2Srk8eN+M7zv1SI6F8N4
+nZRzxt89zZXTXgm3D9SiCipmsgHribqzQLhOaXo4mlSe+nIeLfGETWX54ZjEf7Dd
+sumWXrWIGq3my5Ep7zS68kjZJmpunm5T1qo6grrySOUhDmK5RfSs3SSb4rHKn/19
+gLNGXQKBgAWgUHiwSNFPLLre4i3NWMMqaxIgQPhxLVgNwQErJnS7YSFmlEVynsEz
+ucEO74emp+9lhjInRYpLlyZF1HgpblAQEOWwe1rzpr812kYLY1fflArOljr/7aq1
+Pn9qlt81d+uJFbIJqFFw6yT10L8zZP3a1vAvzatetaje4AzZ7aQz
+-----END RSA PRIVATE KEY-----
