+++ src/gurubase-backend/backend/integrations/bots/github/urls.py
@@ -1,0 +1,14 @@
+
+from django.conf import settings
+from django.urls import path
+from integrations.bots.github.views import github_webhook
+urlpatterns = []
+
+urlpatterns += [
+    path('github/', github_webhook, name='github_webhook'),
+]
+
+if settings.ENV == 'selfhosted':
+    urlpatterns += [
+        path('api/github/', github_webhook, name='github_webhook'),
+    ]
