+++ src/gurubase-backend/backend/integrations/bots/helpers.py
