+++ src/gurubase-backend/backend/integrations/ingestion/jira/jira_strategy.py
@@ -1,7 +1,7 @@
+from integrations.bots.helpers import IntegrationError
+from integrations.models import Integration
+from integrations.strategy import IntegrationStrategy
 from core.exceptions import JiraAuthenticationError, JiraError, JiraInvalidDomainError
-from .strategy import IntegrationStrategy
-from .helpers import IntegrationError
-from core.models import Integration, GuruType
 from core.requester import JiraRequester # Import JiraRequester
 import logging
