+++ src/gurubase-frontend/src/app/actions.js
@@ -801,7 +801,7 @@ export async function getApiKeys() {
 export async function getIntegrationDetails(guruType, integrationType) {
   try {
     const response = await makeAuthenticatedRequest(
-      `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/${guruType}/integrations/${integrationType}/`,
+      `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/integrations/${guruType}/${integrationType}/`,
       {
         method: "GET",
         headers: { "Content-Type": "application/json" }
@@ -884,7 +884,7 @@ export async function deleteApiKey(formData) {
 export async function getIntegrationChannels(guruType, integrationType) {
   try {
     const response = await makeAuthenticatedRequest(
-      `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/${guruType}/integrations/${integrationType}/channels/`,
+      `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/integrations/${guruType}/${integrationType}/channels/`,
       {
         method: "GET",
         headers: { "Content-Type": "application/json" }
@@ -919,7 +919,7 @@ export async function saveIntegrationChannels(
 ) {
   try {
     const response = await makeAuthenticatedRequest(
-      `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/${guruType}/integrations/${integrationType}/channels/`,
+      `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/integrations/${guruType}/${integrationType}/channels/`,
       {
         method: "POST",
         headers: { "Content-Type": "application/json" },
@@ -1017,7 +1017,7 @@ export async function sendIntegrationTestMessage(integrationId, channelId) {
 export async function deleteIntegration(guruType, integrationType) {
   try {
     const response = await makeAuthenticatedRequest(
-      `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/${guruType}/integrations/${integrationType}/`,
+      `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/integrations/${guruType}/${integrationType}/`,
       {
         method: "DELETE",
         headers: { "Content-Type": "application/json" }
@@ -1048,7 +1048,7 @@ export async function deleteIntegration(guruType, integrationType) {
 export async function getIntegrationsList(guruType) {
   try {
     const response = await makeAuthenticatedRequest(
-      `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/${guruType}/integrations/`,
+      `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/integrations/${guruType}/`,
       {
         method: "GET",
         headers: { "Content-Type": "application/json" }
@@ -1082,7 +1082,7 @@ export async function createSelfhostedIntegration(
 ) {
   try {
     const response = await makeAuthenticatedRequest(
-      `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/${guruType}/integrations/${integrationType}/`,
+      `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/integrations/${guruType}/${integrationType}/`,
       {
         method: "POST",
         headers: { "Content-Type": "application/json" },
@@ -1223,7 +1223,7 @@ export async function parseSitemapUrls(sitemapUrl) {
 export async function fetchJiraIssues(integrationId, jqlQuery) {
   try {
     // Construct the endpoint URL using the integrationId
-    const endpointUrl = `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/jira/issues/${integrationId}/`;
+    const endpointUrl = `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/integrations/jira/issues/${integrationId}/`;
 
     const response = await makeAuthenticatedRequest(endpointUrl, {
       method: "POST",
@@ -1262,7 +1262,7 @@ export async function fetchJiraIssues(integrationId, jqlQuery) {
 export async function fetchConfluencePages(integrationId, searchQuery) {
   try {
     // Construct the endpoint URL using the integrationId
-    const endpointUrl = `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/confluence/pages/${integrationId}/`;
+    const endpointUrl = `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/integrations/confluence/pages/${integrationId}/`;
 
     const response = await makeAuthenticatedRequest(endpointUrl, {
       method: "POST",
@@ -1301,7 +1301,7 @@ export async function fetchConfluencePages(integrationId, searchQuery) {
 export async function fetchZendeskTickets(integrationId) {
   try {
     // Construct the endpoint URL using the integrationId
-    const endpointUrl = `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/zendesk/tickets/${integrationId}/`;
+    const endpointUrl = `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/integrations/zendesk/tickets/${integrationId}/`;
 
     const response = await makeAuthenticatedRequest(endpointUrl, {
       method: "GET" // Assuming GET request as no body is needed
@@ -1337,7 +1337,7 @@ export async function fetchZendeskTickets(integrationId) {
 export async function fetchZendeskArticles(integrationId) {
   try {
     // Construct the endpoint URL using the integrationId
-    const endpointUrl = `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/zendesk/articles/${integrationId}/`;
+    const endpointUrl = `${process.env.NEXT_PUBLIC_BACKEND_FETCH_URL}/integrations/zendesk/articles/${integrationId}/`;
 
     const response = await makeAuthenticatedRequest(endpointUrl, {
       method: "GET" // Assuming GET request as no body is needed
